+++ sdk/ai/Azure.AI.Projects.OpenAI/src/Custom/AgentWorkflowPreviewActionResponseItem.cs
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#nullable disable
+
+namespace Azure.AI.Projects.OpenAI;
+
+[CodeGenType("WorkflowActionOutputItem")]
+public partial class AgentWorkflowPreviewActionResponseItem
+{
+    [CodeGenMember("Status")]
+    public AgentWorkflowPreviewActionStatus? Status { get; }
+}