+++ sdk/ai/Azure.AI.Projects.OpenAI/src/Custom/Internal/InternalOpenAIConversationsClient.cs
@@ -3,13 +3,8 @@
 
 #nullable disable
 
-using System;
 using System.Collections.Generic;
-using System.ClientModel.Primitives;
-using System.Text.Json;
 using System.Threading;
-
-using OpenAI;
 
 namespace Azure.AI.Projects.OpenAI;