+++ sdk/ai/Azure.AI.Projects.OpenAI/tests/Samples/Sample_OpenAPI_ProjectConnection.cs
@@ -43,8 +43,8 @@ public class Sample_OpenAPIProjectConnection : ProjectsOpenAITestBase
         AIProjectConnection tripadvisorConnection = await projectClient.Connections.GetConnectionAsync("tripadvisor");
         OpenAPIFunctionDefinition toolDefinition = new(
             name: "tripadvisor",
-            spec: BinaryData.FromBytes(BinaryData.FromBytes(File.ReadAllBytes(filePath))),
-            auth:  new OpenAPIProjectConnectionAuthenticationDetails(new OpenAPIProjectConnectionSecurityScheme(
+            specificationBytes: BinaryData.FromBytes(File.ReadAllBytes(filePath)),
+            authentication:  new OpenAPIProjectConnectionAuthenticationDetails(new OpenAPIProjectConnectionSecurityScheme(
                 projectConnectionId: tripadvisorConnection.Id
             ))
         );
@@ -98,8 +98,8 @@ public class Sample_OpenAPIProjectConnection : ProjectsOpenAITestBase
         AIProjectConnection tripadvisorConnection = projectClient.Connections.GetConnection("tripadvisor");
         OpenAPIFunctionDefinition toolDefinition = new(
             name: "tripadvisor",
-            spec: BinaryData.FromBytes(BinaryData.FromBytes(File.ReadAllBytes(filePath))),
-            auth: new OpenAPIProjectConnectionAuthenticationDetails(new OpenAPIProjectConnectionSecurityScheme(
+            specificationBytes: BinaryData.FromBytes(File.ReadAllBytes(filePath)),
+            authentication: new OpenAPIProjectConnectionAuthenticationDetails(new OpenAPIProjectConnectionSecurityScheme(
                 projectConnectionId: tripadvisorConnection.Id
             ))
         );